+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Airbrake Ruby Changelog
 
 ### master
 
+* Ignored `SystemExit` in the `at_exit` hook, which has fixed the Rake
+  integration for the [airbrake gem][airbrake-gem] gem
+  ([#14](https://github.com/airbrake/airbrake-ruby/pull/14))
+
 ### [v1.0.1][v1.0.1] (December 22, 2015)
 
 * Fixed the `Airbrake.add_filter` block API
@@ -17,6 +21,7 @@ Airbrake Ruby Changelog
 
 * Initial release
 
+[airbrake-gem]: https://github.com/airbrake/airbrake
 [v1.0.0.rc.1]: https://github.com/airbrake/airbrake-ruby/releases/tag/v1.0.0.rc.1
 [v1.0.0]: https://github.com/airbrake/airbrake-ruby/releases/tag/v1.0.0
 [v1.0.1]: https://github.com/airbrake/airbrake-ruby/releases/tag/v1.0.1